fixing bug and updating tests
The bug being fixed manifested when perl was built when
CC was gcc with as a path, usually as a full path to gcc,
e.g., "/path/to/gcc".

Support for full path gcc was added to _update_supported.

Added tests for different perls and different gccs to ensure
the fix; fixed the Windows test and removed linux tests from
ci.yml.

Darwin tests are still breaking.